abis: add socket, netinet/in and fcntl abi to abi-bits
