+++ abis/mlibc/fcntl.h
@@ -1,0 +1,25 @@
+#ifndef _ABIBITS_FCNTL_H
+#define _ABIBITS_FCNTL_H
+
+// constants for fcntl()'s command argument
+#define F_DUPFD 1
+#define F_DUPFD_CLOEXEC 2
+#define F_GETFD 3
+#define F_SETFD 4
+#define F_GETFL 5
+#define F_SETFL 6
+#define F_GETLK 7
+#define F_SETLK 8
+#define F_SETLKW 9
+#define F_GETOWN 10
+#define F_SETOWN 11
+
+// constants for struct flock's l_type member
+#define F_RDLCK 1
+#define F_UNLCK 2
+#define F_WRLCK 3
+
+// constants for fcntl()'s additional argument of F_GETFD and F_SETFD
+#define FD_CLOEXEC 1
+
+#endif // _ABITBITS_FCNTL_H
