+++ options/linux/include/linux/sched.h
@@ -6,8 +6,6 @@
 extern "C" {
 #endif
 
-#include <bits/posix/sa_family_t.h>
-
 #define SCHED_NORMAL 0
 #define SCHED_FIFO 1
 #define SCHED_RR 2
