--- options/posix/include/bits/posix/sa_family_t.h
@@ -1,6 +1,0 @@
-#ifndef MLIBC_SA_FAMILY_T_H
-#define MLIBC_SA_FAMILY_T_H
-
-typedef unsigned int sa_family_t;
-
-#endif // MLIBC_SA_FAMILY_T_H
