+++ options/posix/include/sys/un.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-#include <bits/posix/sa_family_t.h>
+#include <abi-bits/socket.h>
 
 struct sockaddr_un {
 	sa_family_t sun_family;
