+++ sysdeps/lemon/include/abi-bits/socket.h
@@ -1,0 +1,1 @@
+../../../../abis/mlibc/socket.h
