+++ sysdeps/linux/include/abi-bits/fcntl.h
@@ -1,0 +1,1 @@
+../../../../abis/linux/x86_64/fcntl.h
