+++ sysdeps/linux/include/abi-bits/socket.h
@@ -1,0 +1,1 @@
+../../../../abis/linux/x86_64/socket.h
