+++ sysdeps/qword/include/abi-bits/fcntl.h
@@ -1,0 +1,1 @@
+../../../../abis/mlibc/fcntl.h
