+++ sysdeps/qword/meson.build
@@ -17,8 +17,11 @@ if not no_headers
 		'include/abi-bits/seek-whence.h',
 		'include/abi-bits/vm-flags.h',
 		'include/abi-bits/errno.h',
+		'include/abi-bits/fcntl.h',
+		'include/abi-bits/in.h',
 		'include/abi-bits/stat.h',
 		'include/abi-bits/signal.h',
+		'include/abi-bits/socket.h',
 		subdir: 'abi-bits'
 	)
 	install_headers(
