+++ sysdeps/sigma/meson.build
@@ -25,8 +25,11 @@ if not no_headers
 		'include/abi-bits/auxv.h',
 		'include/abi-bits/vm-flags.h',
 		'include/abi-bits/errno.h',
+		'include/abi-bits/fcntl.h',
+		'include/abi-bits/in.h',
 		'include/abi-bits/stat.h',
 		'include/abi-bits/signal.h',
+		'include/abi-bits/socket.h',
 		subdir: 'abi-bits'
 	)
 endif
